refactor: change data type

+++ src/main/java/com/hmwcs/snowflake/config/SnowflakeConfig.java
@@ -14,27 +14,27 @@ public class SnowflakeConfig {
     /**
      * The number of bits allocated for the data center ID in the Snowflake ID.
      */
-    public static final byte DATA_CENTER_ID_BITS = 5;
+    public static final int DATA_CENTER_ID_BITS = 5;
 
     /**
      * The number of bits allocated for the machine ID in the Snowflake ID.
      */
-    public static final byte MACHINE_ID_BITS = 5;
+    public static final int MACHINE_ID_BITS = 5;
 
     /**
      * The maximum value for the data center ID, calculated based on the number of bits allocated.
      */
-    public static final long MAX_DATA_CENTER_ID = ~(-1L << DATA_CENTER_ID_BITS);
+    public static final int MAX_DATA_CENTER_ID = ~(-1 << DATA_CENTER_ID_BITS);
 
     /**
      * The maximum value for the machine ID, calculated based on the number of bits allocated.
      */
-    public static final long MAX_MACHINE_ID = ~(-1L << MACHINE_ID_BITS);
+    public static final int MAX_MACHINE_ID = ~(-1 << MACHINE_ID_BITS);
 
     /**
      * The number of bits allocated for the sequence number in the Snowflake ID.
      */
-    public static final long SEQUENCE_BITS = 12L;
+    public static final int SEQUENCE_BITS = 12;
 
     /**
      * The number of bits to shift the machine ID to the left in the final Snowflake ID.
